requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
attrs==24.3.0
feedparser==6.0.11
frozenlist==1.5.0
idna==3.10
multidict==6.1.0
pillow==11.1.0
propcache==0.2.1
py-cord==2.6.1
qrcode==8.0
sgmllib3k==1.0.0
yarl==1.18.3